Title: 실수 입력받아 계산 2, Time: 11ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240329/실수 입력받아 계산 2/calculation-by-inputting-a-real-number-2.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main() {
     // 여기에 코드를 작성해주세요.
-    int a = 0;
+    double a = 0;
     cout << fixed;
     cout.precision(2);
     cin >> a ;
